vayez.api/src/mail/mail.service.ts
import * as nodemailer from 'nodemailer';
import { Injectable, Logger } from '@nestjs/common';

@Injectable()
export class MailService {
  private transporter: nodemailer.Transporter;
  private readonly logger = new Logger(MailService.name);

  constructor() {
    this.createTestAccount();
  }

  private async createTestAccount() {
    try {
      const testAccount = await nodemailer.createTestAccount();
      this.transporter = nodemailer.createTransport({
        host: 'smtp.ethereal.email',
        port: 587,
        secure: false, 
        auth: {
          user: testAccount.user, 
          pass: testAccount.pass, 
        },
        logger: true,
        debug: true,  
      });
      this.logger.log(`Ethereal account created - User: ${testAccount.user}, Pass: ${testAccount.pass}`);
    } catch (error) {
      this.logger.error('Failed to create test account', error);
    }
  }

  async sendActivationEmail(to: string, token: string) {
    const activationLink = `http://localhost/activate?token=${token}`;
    const mailOptions = {
      from: 'Auth-backend service',
      to: to,
      subject: 'Account Activation',
      html: `<p>Click the link below to activate your account:</p><p><a href="${activationLink}">Activate</a></p>`,
    };
    await this.transporter.sendMail(mailOptions);
  }

  async sendPasswordResetEmail(to: string, token: string) {
    const resetLink = `http://localhost/reset-password?token=${token}`;
    const mailOptions = {
      from: 'Auth-backend service',
      to: to,
      subject: 'Password Reset Request',
      html: `<p>You requested a password reset. Click the link below to reset your password:</p><p><a href="${resetLink}">Reset Password</a></p>`,
    };

    await this.transporter.sendMail(mailOptions);
  }
}
